Add 'refresh', 'pause', and 'resume' commands to hack/dind-cluster.sh
The new refresh command will redeploy the binaries to a dind cluster
and restart all openshift processes that use the binaries so that you
can more easily test changes to openshift in a dind environment.

The pause command will stop the running containers in the cluster, but
not remove them.

The resume command will start containers in the cluster if they were
previosly paused, or otherwise stopped (perhaps from a reboot).